ref: Reorganize Day 5 files
