+++ lib/puzzles/2023/day05/day05.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+require_relative "part1"
+require_relative "part2"
+
+module AdventOfCode
+  module Puzzles2023
+    ##
+    # Advent of Code 2023 - Day 5
+    # https://adventofcode.com/2023/day/5
+    # {include:file:lib/puzzles/2023/day05/README.md}
+    module Day05
+    end
+  end
+end
