+++ lib/puzzles/2023/day05/input.txt
